pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "python-nmap>=0.7.1",
    "streamlit-folium>=0.25.2",
    "streamlit>=1.50.0",
    "trafilatura>=2.0.0",
    "urllib3>=2.5.0",
    "stripe>=12.5.1",
    "requests>=2.32.5",
    "pyjwt>=2.10.1",
    "plotly>=6.3.0",
    "folium>=0.20.0",
    "reportlab>=4.4.4",
    "argon2-cffi>=25.1.0",
]
